move config to lib
