+++ app/AuthProvider/AuthProvider.tsx
@@ -1,6 +1,5 @@
 import { Auth0Provider } from '@auth0/auth0-react';
-
-import { config } from '../config/config';
+import { config } from '@camp/config';
 
 interface Props {
   children: JSX.Element;
