+++ app/domain/index.ts
@@ -1,3 +1,4 @@
 export * from './ClientCookie';
-export * from './Config';
+export * from './Path';
 export * from './ServerErrors';
+export * from './Url';
