+++ cypress/e2e/createProject.cy.ts
@@ -1,8 +1,8 @@
-import { AppRoute } from '../../src/AppRoutes';
-import { createProjectButtonId } from '../../src/components/CreateProject/ProjectButton/CreateProjectButton.ids';
-import { createProjectFormIds } from '../../src/components/CreateProject/ProjectForm/CreateProjectForm.ids';
-import { dashboardHeaderId } from '../../src/components/organisms/Header/DashboardHeader.ids';
-import { navLinkIds } from '../../src/components/organisms/Sidebar/Sidebar.ids';
+import { createProjectButtonId } from '../../app/components/CreateProject/ProjectButton/CreateProjectButton.ids';
+import { createProjectFormIds } from '../../app/components/CreateProject/ProjectForm/CreateProjectForm.ids';
+import { dashboardHeaderId } from '../../app/components/organisms/Header/DashboardHeader.ids';
+import { navLinkIds } from '../../app/components/organisms/Sidebar/Sidebar.ids';
+import { AppRoute } from '../../libs/router/AppRoutes';
 import { admin } from '../fixtures/admin';
 import { projectFixture } from '../fixtures/family';
