+++ cypress/support/commands/visit.command.ts
@@ -1,5 +1,5 @@
-import type { AppRoute } from '../../../src/AppRoutes';
-import { buildUrl } from '../../../src/router/buildUrl';
+import type { AppRoute } from '../../../libs/router/AppRoutes';
+import { buildUrl } from '../../../libs/router/buildUrl';
 
 Cypress.Commands.overwrite('visit', (visit, url, options) => {
   if (options?.params == null) return visit(url, options);
