+++ libs/config/README.md
@@ -1,0 +1,10 @@
+# Config
+
+## Concerns
+
+* Environment Resolution
+* Configuration and Environment Parsing anv Validation
+
+## Dependency Rules
+
+* This modules must be 0 dependency.
