+++ libs/data-layer/README.md
@@ -1,0 +1,15 @@
+# Data-Layer
+
+## Concerns
+
+* Apollo Client Provider
+* Query and Mutation hooks
+* Caching
+* Mapping API types to Client types
+
+
+## Dependency Rules
+
+* @camp/config
+* @camp/domain
+* It should not expose any API types
