+++ libs/data-layer/operations/mutations/CreateFamily.mutation.graphql
