+++ libs/data-layer/operations/mutations/CreateProject.mutation.graphql
